refactor(drop): split mod.rs into submodules
